Migrated from bespoke test iterator to npm module for cartesian products
