+++ package.json
@@ -31,7 +31,8 @@
     "singleQuote": true
   },
   "dependencies": {
-    "option": "^0.2.4",
+    "@stdlib/array-base-unitspace": "^0.2.2",
+    "cartesian-product-generator": "^1.1.1",
     "random": "^5.1.1",
     "winston": "^3.17.0"
   }
